* tracker/4.14/gse:
  hyper_dmabuf: fix map failure issue when assign 4G memory to UOS (v2)
  hyper_dmabuf/virtio: Handle S3 resume correctly (v2)
  hyper_dmabuf/virtio: Add support for VBS_RESET_DEV ioctl (v2)
  hyper_dmabuf/virtio: bugfix on acrn_ioreq_add_iorange() usage
  hyper_dmabuf/virtio: Correctly cleanup front end connections
  hyper_dmabuf/virtio: Protect virtqueue operations with spinlock.

+++ drivers/dma-buf/hyper_dmabuf/virtio/hyper_dmabuf_virtio_shm.h
@@ -25,7 +25,7 @@
 #ifndef __HYPER_DMABUF_VIRTIO_SHM_H__
 #define __HYPER_DMABUF_VIRTIO_SHM_H__
 
-int virtio_share_pages(struct page **pages,
+long virtio_share_pages(struct page **pages,
 		       int domid, int nents,
 		       void **refs_info);
